requirements.txt
appdirs==1.4.4
attrs==20.3.0
autocensus==2.0.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
Fiona==1.8.18
geomet==0.2.1.post1
geopandas==0.8.1
h11==0.11.0
httpcore==0.12.1
httpx==0.16.1
idna==2.10
jellyfish==0.6.1
multidict==5.0.2
munch==2.5.0
mypy-extensions==0.4.3
nest-asyncio==1.4.3
numpy==1.19.4
packaging==20.4
pandas==1.1.4
pandera==0.5.0
pyparsing==2.4.7
pyproj==3.0.0.post1
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
rfc3986==1.4.0
Shapely==1.7.1
six==1.15.0
sniffio==1.2.0
socrata-py==1.0.5
tenacity==6.2.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
us==2.0.2
wrapt==1.12.1
yarl==1.6.3
